Add games_scummed to LeaderboardBaseFields
Forgot that we track this for a trophy, and it's used in enough places
that awkwardly querying games everywhere it's needed is inefficient, so
some denormalization is in order (like for the rest of the aggregated
fields).

We're not planning on having an actual leaderboard for this, though...
uh... right?