Fix: Gracefully handle "orphaned user" in clan management
I would be surprised if this comes up during the actual tournament, but
it happened during testing: I wiped the db to re-apply migrations, but
this didn't wipe Django's built in user table. So I was still able to
log in and access my clan management page, which promptly errored when I
tried to do anything with it.

This adds a fallback where if a user logs in and doesn't have an id
match on Player, it will look for a matching name and link the user and
Player together if found. If there's a logged in user who doesn't match
any Player's name in the database at all, then no. Fail. This is the
first case where we have to return an actual error to the user. (Again,
I don't see this case happening in production because we wouldn't be
casually wiping the Player table.)