Add partial leaderboards page at /leaderboards
The main new big feature of 2021 is leaderboards. Given their
prominence, they get a whole page linked from the top (replacing
Trophies... but we still have trophies, so we'll need to figure out some
way to navigate to those.)

All leaderboards are loaded on the page, but only one is shown at a
time (because they could get quite long with all the players/clans).
They're currently changed via a combo box with a bit of JS bound to it.
I kind of want to add at least left/right arrow key navigation as well.

As of this commit, only 3 leaderboards are functional, but others are
expected to follow easily along the same pattern.

Also adds a shorthand for querying only ascended players via a Manager
class (Player.ascended.*).