Rearrange title bar
The "Home" page is dropped to make room for having both Leaderboards and
Trophies pages. The / endpoint is still accessible by clicking on the
logo and now also by the big page title.

The rearrangement of the links in the title bar probably isn't perfect,
but it balances the width of the left and right text pretty well.