Add a Game convenience method for role-race-gender-align string
This was used in two places now, which is one too many, so put the logic
in one place. (It will also be used in streak computation).

Additionally, this is useful because it makes it less likely someone
will go and use the Game gender and align fields rather than gender0 and
align0 which the scoreboard actually uses.