Rework the layout of single-player Ascensions and Recent Games
In testing, I created an ascension with every possible conduct and
discovered that it expanded the Ascensions table across the entire
screen, dragging the Recent Games table with it. This looked bad. So
this introduces a new rule for top-level tables to not exceed 70% of
page width (same as some other tables) and disconnects the Ascensions
table from Recent Games, splitting them into their own tables.