Fully implement the leaderboards page
This is what will likely be the final 2021 design: a dual leaderboard
(player/clan) for each category. The displayed one can be changed with
the combo box, arrows next to the combo box for next/prev, or arrow
keys.

Rendering the leaderboards page is VERY inefficient database-wise (using
the swanky new db query counter, almost 150 hits to render the page with
only a couple of players and clans). I tried many ways of optimizing it
more, but wasn't able to get there and with Nov 1 looming decided an
inefficient working product is better than an optimized broken one. K2
is optimistic that the server can handle it.

Also remove AscendingPlayersManager which is not used anymore.