Add unique deaths list to single player/clan pages
This is not as feature complete as in the previous scoreboard where, for
clans, the achieving player and time of each death are listed. That's
for another year. On the other hand, the previous scoreboard didn't show
unique deaths for individual players (but here, it's a player
leaderboard competition so we really ought to).