Implement a significant optimization in leaderboards
This is a huge shift in terms of the number of queries produced in this
view - down to 2 from almost 900 (and more significantly, down to O(1)
from O(#clans + #players).) The page load speed difference is obvious
even on localhost.

A large part of this was extra queries to look up a Game and allow it to
hit the Player and Source tables in order to generate its dumplog. Very
clean code-wise. Very bad database-wise. Now, the 2 queries fetch all
needed data to construct the dumplog on the backend. (Games still do
this in other views. The plan is to phase those out.)

This also overhauls the interaction between LeaderboardsView and its
template - now the view runs the show and provides all the data for
the leaderboards. Because the view is now providing a well-formatted
dict/list structure, we can get rid of the ugly get_dict_field and
its friends, and I deleted those.