Limit achievement table max width
Some new, super-long achievement descriptions were messing with the
layout of the player page.  Add a maximum width for cells in the
achievement table, so that very long descriptions will wrap over to the
next line instead of making the table extremely wide.  Should only
cause two or three descriptions to wrap, the vast majority are still
short enough to fit in one line.