Add the ability to pollxlogs from a local file
This is not intended for production at all but is mainly useful for dev
purposes, when one wants to load a custom-edited xlogfile (say, one
containing a lot of streaks to test the streak rules) without going
through the normal pollxlogs process of pulling from a server.

The code is not as clean as I would like but I prefer it to having two
duplicate "loop over an xlogfile and create Games from it" functions.

Also improve message printing from pollxlogs - it will print a total
number of new Games created at the end, and any Game which fails to be
created (such as one outside the tournament time) will print an error
message, so I don't have to go through this process of figuring out why
games weren't getting created next year.

Also add documentation to README and reset_db.sh clarifying some things
and processes I found confusing upon returning this year.