Remove GameManager and move xlog-to-Game conversion to pollxlogs.py
This addresses a TODO from last year - to recheck whether GameManager
was actually doing anything that a Manager class is intended for. I
looked at the documentation and no, it was not. All it was doing was
serving as the place where from_xlog lived.

So this moves the logic of translating an xlog entry into a Game into
pollxlogs, and gets rid of the manager class.