Rename Never Kill Foo trophies to Keep Foo Alive
To clear up confusion with people thinking it meant as long as you did
not kill the monster personally, it would qualify for the trophy (which
is not the case; they have to not die at all).

Also update the descriptions to clarify that you only have to do this in
one game to qualify for the trophy, not do it across all
ascensions/games.