Add support for achievements earned in an in-progress game
This is the web server half of the new feature for showing achievements
that will be earned as soon as a game finishes. It relies on the tnnt
binary creating and making available files with a certain format that
record the player's name and the achieve and tnntachieve fields.

"Temp achievements" as they are mostly called here are populated from
these files during aggregation, and the clan and player web pages will
display them with a yellow symbol to indicate an state in between
achieving them now and not having them yet. They are not treated as
earned achievements, and do not count towards totals. (We could make
things such that if you have a temp achievement attributed to you in one
of these files, the scoreboard just gives you the achievement, but it's
probably better to stick with the model of it not officially counting
until the game is in the xlogfile).