Fix Near Death Experience small caps method
Instead of targeting it by the exact in-game achievement number (which
can change whenever new achievements are added, requiring the CSS
stylesheet to be manually updated), target it by the define
(A_TALKED_TO_DEATH) so it will always be correct regardless of the order
of the list.

I made up some fake defines for the vanilla achievements, not sure if
there is anything more appropriate to use (like the u.uachieve.foo field
names?).

Also I noticed some of the vanilla achievements were misnumbered, so fix
those.