Add swap chest donations leaderboard
This uses the new donor-tracking functionality in the game binary,
records, and reports it. This is under the assumption that the donors
list is published somewhere freely accessible by HTTP, rather than
relying on the server operator to sync the donors file to a certain
local location. During aggregation the scoreboard will download and
parse these lists and credit players with 1 donation for each time they
appear there.