Track the Z-score of players and clans and add a leaderboard for it
After a bit of discussion, we decided to go with role-based Z-score
rather than role+race based, or even role+race+align based, on the
grounds that it is less interesting to award full credit to a player
doing slightly different combinations of the same role, and would tend
towards people's Z-scores closely approximating their number of
ascensions. Plus, basing it on role only is the traditional definition
of Z-score anyway.