Make invite acceptance button explicit
Or try to, at least, because I'm not sure how to test locally to see
exactly how this will look.  It could need some CSS to tweak the
padding between it and the "accept invite" or something like that.

I clicked on one of the clan names earlier thinking it would show me the
clan page (and I guess there would be an 'accept invite' button on that
page?) and was surprised that it immediately accepted the invite.  Maybe
that's just me but hopefully this should clear up any confusion, no
matter how unlikely.