Combine temp achievements across servers
Not wiping the temp achievements status for the player before each file
should hopefully mean that the site will display (FILE_US | FILE_EU)
if multiple games are occuring simultaneously, instead of whichever is
the last file on the list overriding all the others.  I'm not sure it
will work that way because I'm not sure how temp_achievements.add reacts
to being given the same achievement multiple times, but based on the
Django docs for many-to-many relationships[0] I think it should be OK.
If it causes problems we can revert this.

[0] https://docs.djangoproject.com/en/4.2/topics/db/examples/many_to_many/