Validate input for utf8mb3 compatibility
This is inspired by suggestions in GitHub issue #25 by user ciraben as a
way to avoid producing some very involved MySQL errors with full
backtraces when someone tries to input a character which cannot be
represented in the 'utf8mb3' encoding (which the MySQL database
currently uses).  Try to identify characters that will cause that error
and throw a simplified error message rather than feeding them into the
db.  Hopefully this will do the trick.

After this year's tournament we can try to convert the database to
'utf8mb4' again, and hopefully can figure out a way that avoids
emoji-related errors that continued to crop up when that text encoding
was tried previously.  Once that is successful this could be removed or
greatly narrowed in scope (if there is a subset of 4-byte characters
that ought to remain disallowed).