Update TNNT svg logo; notice of shirt sales on header.
This is qt's code, but wasn't officially committed.

The svg logo now uses the correct font for the @ symbol, and has a
shadowing effect like the rest of the text has in the header.

The notice of the tshirt sales, that's not always going to be a TNNT
fixure, but it will be for the foreseeable future. Easy enough to
comment out the bit in the header file when not needed.