Fix Lesser Role/Race check
crunchbubba wrote an xlogfile parsing script to monitor clan progress
towards various Lesser Role trophies, and reported that Surly Dinos had
received "Lesser Priest" without having actually achieved all the
requirements.  Turns out the back end was checking if either of the bits
0x400 (Sokoban) or 0x200 (Mines' End) were set (becuase it was just
looking for a nonzero result from the bitwise AND), but the Lesser Role
and Lesser Race trophies which the 'mines_soko' field is used for are
meant to require both accomplishments.