Close two more TODOs
One of these was to check if it would be possible for the points
reported in an xlog (a long type in the TNNT source) to overflow a
Python int type (in case it is temporarily read into an int). But this
is python 3 and I'm pretty sure it handles larger-than-32-bit integers
seamlessly. So confirmed, there is no need to do anything special for
that.

The other one was to rename 'points' to 'score' on the scoreboard side.
Since the xlogfile uses 'points' I don't see much benefit in doing that.