Change Highest Scoring Ascension to Highest Scoring Game
Taking care of one of the many "post 2021 TODOs" (how convenient that I
called them that and not 2022 TODOs). As noted in the comment, at the
upper extremes of points, it doesn't really matter whether you ascended
the game or not.

Only minor changes were needed to pull this from the set of all games by
a player/clan instead of only the set of their wins.