Resolve another old TODO (dumplog link efficiency) by withdrawing it
We could change this to store a dumplog URL with every Game, resulting
in a lot more storage bloat but a faster and more straightforward way of
getting a dumplog link with a game, but after discussion I am going to
stick it in the "we'll do something about it when it becomes a problem".
The only place a Game needs to do a three-table join to get a dumplog
link, again and again for a bunch of Games, is when rendering player
streaks, and in that case there aren't that many Games. So, change the
TODO to a comment explaining this.

Also remove the TODO from dumplog_utils.py because resolving it was
predicated on storing a dumplog link on Games.