Fix: server threw a bunch of errors when failing to connect to dgl db
The logging statements were misformatted, not containing any %s
specifier for the string parameter. I think this was never a problem
because in production it's been able to connect to the real dgl db, and
normally on a dev build you change it to the local db in settings.py.
Except I didn't, and it wasn't able to find anything at the real dgl db
path, exposing this error.