Resolve another TODO (anchor links on singleplayerorclan page)
Scrolling around to see your achievements or unique deaths is annoying.
I don't know how many people view the scoreboard on a mobile browser,
but it would be even more annoying for anyone who does.

Fortunately, I left a TODO to add anchor links to the page, so I went
ahead and did that. Rather than trying to add a separate list of links,
the links to the sections below the stats summary just use the stats
themselves. If a section doesn't appear on the page (i.e. there is no
streaks section because you don't have any streaks), the link won't
appear either.