Resolve a few old TODOs by implementing clan messages
A deferred feature from the 2021 rewrite. We actually already had a clan
message field in the model, but it was unused. Now it can be used - any
admin of a clan can set the message, and only members of the clan can
read it. The intended use for this is so admins can provide a message
like "We're coordinating efforts on ##ourawesomeclan on Libera, come
join us there" and not have to manually pass along this information to
every member that joins, or even be unable to coordinate with a member
that joins the clan but isn't already in contact with clan members.
There's also less risk of leaking a private channel the clan admins would
rather not disclose.

Of course, you don't HAVE to use the message for coordination. You can
also put an encouraging message, state your goals, or do anything
really (within 512 characters).

Tested to confirm the clan message cannot be used for script injection
attacks.

In the process, I changed the strict disabling of / which previously
applied to all text input forms on the site to apply only to the clan
creation. That technically means you can invite a player with a / in
their name now, but in practice no such player will exist.