Add a little clarification to Lesser Race/Role definition
A player new to either NetHack or the tournament might suppose that a
game only counts for Lesser Race/Role if they get the two prizes... and
also manage to bring them out of the dungeon safely, either escaping or
ascending, and a game where they die and fail to do that doesn't count.

I added a sentence emphasizing the game still counts if you obtain both
prizes but then die afterwards.

While here, also fix another hyphen masquerading as an emdash.