Add Unique Deaths page
This has been something I wanted to do for a while. It summarizes unique
deaths across the tournament, with additional details including the
first player to get a unique death and the time it was achieved.
Figuring out how to query the Django ORM for this was difficult; thanks
to the anonymous friend of a friend who helped me out with it.

Knocks off 3 old TODOs relating to unique deaths.