Add Stats page providing a bunch of overall tournament statistics
Like the Unique Deaths page, this page is not currently reachable from
the main site. That will be changed soon. This new Stats page does,
however, contain links to both the unique deaths page and the previously
unreachable achievements page.

As noted in a comment, other stats can be added; these are only the ones
I thought of now and thought were feasible to implement.