Fix server settings when DEBUG is turned on
The previous commit allowed DEBUG to finally be turned off on the
production server by using non-Django middleware to serve static files,
but at the cost of unconditionally requiring that middleware for all
users of the server, including anyone running it on a dev machine where
extra middleware isn't needed and DEBUG should be on.

This should make the settings work in both cases. The whitenoise
middleware will now be conditionally added in where needed when DEBUG is
off, as well as enabling localhost in the allowed hosts and setting the
login db to the local test copy when DEBUG is on.