Move trophy requirements from hovertext to regular text
I am preparing to pare down the slab of errata describing most of the
trophies. In fact, most of the trophies already have pretty good
descriptions already, but requiring people to hover over a trophy name
to find out the requirements to earn that trophy is pretty terrible from
a UI design perspective. This change does make the trophies take up more
page space, but only as long as not a lot of people have earned them. Once
players and clans start earning trophies, the vertical space necessary
to put their names should surpass that of the trophy name and
description.

Hover text also happens to be impossible to read on mobile, so this is a
benefit for any hypothetical people viewing the site on mobile devices.

I also tweaked the description of each trophy that requires different
starting alignments to explicitly say "starting" - helping prevent
confusion where changing one's alignment to get a new combo won't work.
The FAQ page will cover this too, but it doesn't hurt to have here too.