Change shirt notification to announcement banner
We could conceivably use this large yellow bar at the top for other
announcements at some point, but they will probably not be
shirt-related, so change the name to be less specifically about shirts.
Also change its commented out state from an HTML comment to a Django
template comment so that unused HTML won't be sent over the wire.

Ideally we could make the announcement configurable so that it doesn't
require changing the template and pushing a commit, but the juice
certainly isn't worth the squeeze on that one.