Resolve another TODO: clean out unused CSS rules
I think most of what I deleted was a holdover from the pre-2021
scoreboard and referenced classes and ids that are no longer used
anywhere. Everything I deleted, I checked for the existence of on any
current template.

We might at some point bring back a trophy that is clan- or
player-exclusive, so I left the trophy-na class commented out instead of
deleted.