Revamp TNNT header logo for 2024
The 2023 change to the logo was to address that many viewers of the
website didn't have the original GNOME font it used (Cantarell) and thus
the @ in it was not rendering right, but the fix to a more standard font
can result in a @ in a font that's poorly centered for some viewers (me,
at least). This fixes the problem by changing the original @ shape to a
path, so it does not depend on the viewer's browser text rendering.

In addition, I added a drop shadow in the SVG instead of in CSS, which
was inconsistently rendering as a drop shadow for the whole SVG (the
blue rectangle) when I loaded the page with the development server. The
rule in CSS is no longer needed and I removed it.

Also, saved it as a plain SVG this time so there's none of the Inkscape
metadata (useless to viewers) and cleaned up unused gradients and other
definitions.